Moved trace_operations.py to faultmeshops and made appropriate changes to trace_to_mesh.py.
