+++ src/eq_fault_geom/faultmeshops/trace_operations.py
