+++ src/eq_fault_geom/faultmeshops/trace_to_mesh.py
@@ -1,10 +1,13 @@
+from typing import Union
+
 import geopandas as gpd
 import pandas as pd
 from shapely.geometry import LineString
 import numpy as np
+
 import meshio
-from eq_fault_geom.geomops.trace_operations import calculate_dip_rotation
-from typing import Union
+
+from trace_operations import calculate_dip_rotation
 
 
 def project_mesh_from_trace(fault_info: pd.Series, cell_type: str = "triangle",
